+++ src/Constants.cs
@@ -6,5 +6,6 @@
         public const string FileName = ".editorconfig";
         public const string Homepage = "http://editorconfig.org";
         public const string DefaultFileContent = "[*]\r\nend_of_line = crlf\r\n\r\n[*.xml]\r\nindent_style = space";
+        public const string HelpLink = "https://github.com/madskristensen/EditorConfigLanguage/wiki/Error-codes#{0}";
     }
 }